% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expectations.R
\name{expect_date}
\alias{expect_date}
\title{Check if a set of columns are dates. Useful for debugging PowerBI R scripts with dates.}
\usage{
expect_date(df, cols, stop_if_fail = TRUE, return_df = TRUE)
}
\arguments{
\item{cols}{character vector of columns names to check if dates}

\item{stop_if_fail}{T/F for whether to consider failure an error}

\item{return_df}{T/F whether to end function with dataframe input (as in if a check in part of a pipe)}
}
\description{
Check if a set of columns are dates. Useful for debugging PowerBI R scripts with dates.
}
