% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expectations.R
\name{expect_dimensions_between}
\alias{expect_dimensions_between}
\title{Check number of rows and columns matches expectation of range}
\usage{
expect_dimensions_between(df, min_nrow = 0, max_nrow = Inf,
  min_cols = 0, max_cols = Inf, stop_if_fail = TRUE,
  return_df = TRUE)
}
\arguments{
\item{df}{data frame to check. If a vector is given, it converts it into a one column dataframe to check length as rows}

\item{min_nrow}{Integer of inclusive minimum OK}

\item{max_nrow}{Integer of inclusive maximum OK. Default is same as minimum}

\item{stop_if_fail}{T/F for whether to consider failure an error}

\item{return_df}{T/F whether to end function with dataframe input (as in if a check in part of a pipe)}
}
\description{
Check number of rows and columns matches expectation of range
}
