% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expectations.R
\name{expect_no_duplicates}
\alias{expect_no_duplicates}
\title{Check if a data frame has duplicates in a given column. If a vector is given, check for duplicates in the vector.}
\usage{
expect_no_duplicates(df, test_columns = NA, stop_if_fail = TRUE,
  report_duplicates = TRUE, return_df = TRUE)
}
\arguments{
\item{df}{data frame to check}

\item{test_columns}{character vector name of column expecting no duplicates, if number vector given, column numbers}

\item{stop_if_fail}{T/F for whether to consider failure an error}

\item{report_duplicates}{T/F for whether to return a partial list of the top duplicates if failure}

\item{return_df}{T/F whether to end function with dataframe input (as in if a check in part of a pipe)}
}
\value{
several options depending on whether it fails or succeeeds
}
\description{
Check if a data frame has duplicates in a given column. If a vector is given, check for duplicates in the vector.
}
\examples{
expect_no_duplicates(mtcars, "cyl")
# [1] "top duplicates..."
# A tibble: 3 x 2
# Groups:   cyl [3]
#cyl     n
#<dbl> <int>
#1     4    11
#2     6     7
#3     8    14
# Error in ifelse(stop_if_fail, stop(paste0("Duplicates detected in column: ",  :
                                                Duplicates detected in column: cyl

expect_no_duplicates(rownames(mtcars))
# [1] "no vector duplicates...OK"
}
