% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expectations.R
\name{expect_no_nas}
\alias{expect_no_nas}
\title{Check if there are any NA values in a data frame, or specified column, within a tolerance}
\usage{
expect_no_nas(df, test_column = NA, na_tolerance = 0,
  return_df = TRUE)
}
\arguments{
\item{test_column}{character string for column to test - optional}

\item{na_tolerance}{number of NA allowed before failure, default is zero}

\item{return_df}{T/F whether to end function with dataframe input (as in if a check in part of a pipe)}
}
\description{
Check if there are any NA values in a data frame, or specified column, within a tolerance
}
\examples{
expect_no_nas(mtcars, "cyl")
# [1] "Detected 0 NAs...OK"
expect_no_nas(mtcars)
# [1] "Detected 0 NAs...OK"
expect_no_nas(c(0, 3, 4, 5))
# [1] "Detected 0 NAs...OK"
expect_no_nas(c(0, 3, NA, 5))
# Error in expect_no_nas(c(0, 3, NA, 5)) : Detected 1 NAs
}
