% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expectations.R
\name{expect_values_only_in}
\alias{expect_values_only_in}
\title{Check that values in a discrete or categorical vector are within a set of acceptable values}
\usage{
expect_values_only_in(test_vector, correct_vector)
}
\arguments{
\item{test_vector}{vector to test}

\item{correct_vector}{vector of all acceptable values}
}
\description{
Check that values in a discrete or categorical vector are within a set of acceptable values
}
\examples{
expect_values_only_in(mtcars$cyl, c(2, 4, 6))
# Error in expect_values_only_in(mtcars$cyl, c(2, 4, 6)) :
#   8 value not found in list given
expect_values_only_in(mtcars$cyl, c(2, 4, 6, 8))
# [1] "all values expected...OK"
}
