% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expectations.R
\name{expect_zero_rows}
\alias{expect_zero_rows}
\title{Check if a dataframe has no rows. If a vector is given the length of the vectors is checked.}
\usage{
expect_zero_rows(df1, stop_if_fail = TRUE, report_rowcount = FALSE,
  return_df = TRUE)
}
\arguments{
\item{df1}{dataframe or vector to check}

\item{stop_if_fail}{T/F for whether to consider failure an error}

\item{report_rowcount}{T/F for whether to return the number of rows}

\item{return_df}{T/F whether to end function with dataframe 1 input (as in if a check in part of a pipe)}
}
\value{
several options depending on whether it fails or succeeeds
}
\description{
Check if a dataframe has no rows. If a vector is given the length of the vectors is checked.
}
\examples{
expect_zero_rows(mtcars[mtcars$cyl == 0, ], return_df = F)
# [1] "No rows found as expected...OK"

expect_zero_rows(mtcars$cyl[mtcars$cyl == 0])
# [1] "No rows found as expected...OK"
# numeric(0)

expect_zero_rows(mtcars)
# Error in ifelse(stop_if_fail, stop(paste0("Different number of rows: ",  :
#    Different number of rows: 32 vs: 0
}
