% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expectations.R
\name{expect_same_number_of_rows}
\alias{expect_same_number_of_rows}
\title{Check if a dataframe has the same number of rows as another, or else 0 rows. If vectors are given the lengths of the vectors are compared.}
\usage{
expect_same_number_of_rows(df1, df2 = data.frame(),
  stop_if_fail = TRUE, report_rowcount = FALSE, return_df = TRUE)
}
\arguments{
\item{df1}{dataframe or vector to check (required)}

\item{df2}{optional second dataframe or vector to compare (if not given, defaults to zero row data frame)}

\item{stop_if_fail}{T/F for whether to consider failure an error}

\item{report_rowcount}{T/F for whether to return the number of rows}

\item{return_df}{T/F whether to end function with dataframe 1 input (as in if a check in part of a pipe)}
}
\value{
several options depending on whether it fails or succeeeds
}
\description{
Check if a dataframe has the same number of rows as another, or else 0 rows. If vectors are given the lengths of the vectors are compared.
}
\examples{
expect_same_number_of_rows(mtcars, mtcars)
# [1] "Same number of rows...OK"

expect_same_number_of_rows(mtcars, iris)
# Error in ifelse(stop_if_fail, stop(paste0("Different number of rows: ",  :
#    Different number of rows: 32 vs: 150

expect_same_number_of_rows(mtcars)
# Error in ifelse(stop_if_fail, stop(paste0("Different number of rows: ",  :
#    Different number of rows: 32 vs: 0
}
